requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pymongo==4.7.3
pydantic==2.7.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.0
